requirements.txt
django==1.11
psycopg2
django-admin-list-filter-dropdown
djangorestframework
markdown
django-filter